Split ChatWidget tests into dedicated test files (Plan 00260)
Extract queuing, attachments, interactive questions, streaming scroll, and voice input test suites from ChatWidget.test.tsx into dedicated test files with shared test utilities.
